refactor: Grouped mapFilePaths under utilities for files
