+++ src/utils/files.js
@@ -1,5 +1,6 @@
 export * from './files/copy-files.js';
 export * from './files/create-files.js';
+export * from './files/map-file-paths.js';
 export * from './files/move-files.js';
 export * from './files/remove-directory-if-empty.js';
 export * from './files/remove-files.js';
